requirements.txt
Django==4.1.3
django-glrm
pandas
django-multiselectfield
django-cors-headers
python-dotenv
psycopg2-binary
djangorestframework==3.14.0
markdown
celery[redis]==5.2.7
django-celery-beat==2.4.0
openpyxl
pytest
pytest-django
pytest-mock==3.10.0
mock
model_bakery==1.9.0
Faker
Pillow
 django-phone-field